windows instalation

+++ build/install/config.php
@@ -8,18 +8,25 @@ for($i = 1; $i < $argc; $i++) {
 }
 
 if(empty($filename)) {
-    echo "\n\033[1;31mERROR!\033[0m\n";
+    echo "ERROR!\n";
     echo "Param -f=filename not set it must point to emercoin.conf file\n";
     die();
 }
 
 if(!file_exists($filename)) {
+	$dirname = dirname($filename);
+	if(!file_exists($dirname)) {
+		echo "Directory '$dirname' does not exist !\n";
+		echo "Did you install Emercoin wallet ?";
+		die();
+	}
+	
     touch($filename);
     file_put_contents($filename, "rpcuser=rpcuser
 rpcpassword=rpcpassword
 rpcallowip=127.0.0.1
 rpcport=8332");
-    echo "File \033[1;32m$filename\033[0m created, restart Emercoin service !";
+    echo "File '$filename' created, restart Emercoin service !";
     die();
 }
 
@@ -29,5 +36,5 @@ $content = preg_replace("/(.+?)(#.+?)(\n)/is", '$1$3', $content);
 $config = parse_ini_string($content);
 
 $configfile = 'config.json';
-touch($configfile);
+if(!file_exists($configfile)) touch($configfile);
 file_put_contents($configfile, json_encode($config));
